:recycle: refactor(#168/user): user 컨트롤러 관련 클래스 ninedot-api 모듈로 이동
